fix(typeck): add receiver type check before dyn trait dispatch and unify usize types
- Change:
  - In `lower_method_call_expr`, before using dyn Trait dispatch for a
    method call, verify that the receiver's concrete type name matches
    the `type_name` stored in the dyn_trait_plan (`call.type_name`). If
    the receiver is `Infer` or its type name doesn't match, fall through
    to the "no method found" error path. Prevents incorrect dyn Trait
    dispatch on mismatched types (e.g., calling a method on `i32` when
    the trait impl exists for `bool`).
  - Unify length/size types from `i64` to `usize` across:
    - `str::len` return type in prelude and primitive intrinsic.
    - `String.ptr/len/cap` and `Vec.ptr/len/cap` fields in prelude.
    - All internal `i64` constants and field projections in `intrinsic_lower.rs`
      and `primitive_intrinsics.rs` (Phase A of architecture roadmap).
- Update:
  - `src/mir/lower/expr_variants.rs` – add type_name extraction and comparison
    logic before dyn Trait call generation.
  - `src/stdlib/prelude.rs` – change `String` and `Vec` field types to `usize`;
    adjust `len()` return types and initializer constants.
  - `src/mir/lower/intrinsic_lower.rs` – replace all `i64` types and constants
    with `usize` for length/capacity/index values.
  - `src/mir/lower/primitive_intrinsics.rs` – update `str::len` return type
    and internal len handling to `usize`.
  - `tests/v0/stage18/plan/stage18_296_trait_impl_primitive_tests.rs` – change
    12 negative tests from `assert_eq(0)` to `assert_ne(0)` now that typeck
    correctly rejects calls on types without trait impls.
  - `tests/v0/stage16/plan/stage16_69_assoc_type_driver_tests.rs` – correct
    struct name from `MyWrapper` to `MyBox` to match source.
  - `tests/v0/stage5/plan/driver_dyn_trait_plan_integration_tests.rs` – adjust
    expectations: `Infer` receiver no longer uses dyn Trait dispatch, so count
    of dyn calls drops to 0 in affected tests; add `#[allow(dead_code)]` to
    helper.
  - `docs/worklog.md` – append stage18.297 and stage18.298 entries with
    reasoning, verification, and architecture review results.
  - `README.md` and `RELEASE_NOTES.md` – update version and status descriptions
    to reflect Phase A (`i64→usize`) completion and class-Rust architecture
    corrections.
- New:
  - `docs/develop/v0/stage-18/plan-18.298.md` – deep architecture review
    identifying 7 dimensions of design deviations vs. Rust model, with a
    4-phase roadmap to eliminate special-case solutions (marker bodies,
    intrinsic dispatch, early interception) and achieve a fully class-Rust
    primitive type extension model.
- Refs:
  - README.md
  - RELEASE_NOTES.md
  - docs/develop/v0/stage-18/plan-18.298.md
  - docs/worklog.md
  - src/mir/lower/expr_variants.rs
  - src/mir/lower/intrinsic_lower.rs
  - src/mir/lower/primitive_intrinsics.rs
  - src/stdlib/prelude.rs
  - tests/v0/stage16/plan/stage16_69_assoc_type_driver_tests.rs
  - tests/v0/stage18/plan/stage18_296_trait_impl_primitive_tests.rs
  - tests/v0/stage5/plan/driver_dyn_trait_plan_integration_tests.rs
- Version: v0.493.0

+++ src/stdlib/prelude.rs
@@ -107,7 +107,7 @@ struct Box<T>(*mut T)
 // Per §1.0 原則 6 (通解>特例): one String type — no per-encoding special cases.
 // Per §2 原則 9 (正确>妥协): the &str alias compromise is removed — real
 // owned String is the correct design.
-struct String { ptr: *mut u8, len: i64, cap: i64 }
+struct String { ptr: *mut u8, len: usize, cap: usize }
 // Stage 18.185 (TD-STRING-INTRINSICS): String methods.
 //
 // String::len() — returns the byte length (field access).
@@ -118,8 +118,8 @@ struct String { ptr: *mut u8, len: i64, cap: i64 }
 // Per §1.0 原則 6 (通解>特例): methods defined in prelude source, not
 // intrinsics — reuses existing field access + method resolution.
 impl String {
-    fn len(&self) -> i64 { self.len }
-    fn new() -> String { String { ptr: 0 as *mut u8, len: 0, cap: 0 } }
+    fn len(&self) -> usize { self.len }
+    fn new() -> String { String { ptr: 0 as *mut u8, len: 0usize, cap: 0usize } }
 }
 // Stage 18.195 (TD-VEC-MVP): Vec<T> — owned dynamic array.
 //
@@ -130,13 +130,13 @@ impl String {
 // Per §1.0 原則 6 (通解>特例): one Vec type for all T (generic, not per-type).
 // Per §2 原則 9 (正确>妥协): MVP uses ptr/len/cap layout (not Vec<u8> wrapper
 // like Rust's Vec<T> { buf: RawVec<T>, len }). Simplification acceptable.
-struct Vec<T> { ptr: *mut T, len: i64, cap: i64 }
+struct Vec<T> { ptr: *mut T, len: usize, cap: usize }
 // Stage 18.238 (TD-INTRINSIC-OVERUSE Phase 1): Vec methods via prelude impl.
 // Per §1.0 原則 6 (通解 > 特解): methods defined in prelude source, not
 // hardcoded MIR lower intrinsics. Standard method resolution handles these.
 impl<T> Vec<T> {
-    fn new() -> Vec<T> { Vec { ptr: 0 as *mut T, len: 0, cap: 0 } }
-    fn len(&self) -> i64 { self.len }
+    fn new() -> Vec<T> { Vec { ptr: 0 as *mut T, len: 0usize, cap: 0usize } }
+    fn len(&self) -> usize { self.len }
 }
 // Stage 18.284 (TD-INTRINSIC-OVERUSE Phase 2-A): str primitive methods.
 //
@@ -165,7 +165,7 @@ impl<T> Vec<T> {
 // checker.rs (KNOWN_INTRINSIC_METHODS) — typeck works naturally with
 // the real prelude signatures.
 impl str {
-    fn len(&self) -> i64 { loop {} }
+    fn len(&self) -> usize { loop {} }
     fn is_empty(&self) -> bool { loop {} }
     fn as_bytes(&self) -> &[u8] { loop {} }
 }
